Change variable type that was too small for passwords with more than 5 characters

+++ table.c
@@ -20,13 +20,13 @@ int l;
 long rows;
 unsigned __int128 universe;
 
-void getPwd(char* pwd, long pwdIndex)
+void getPwd(char* pwd, unsigned long long pwdIndex)
 {
-	long value = pwdIndex;
+	unsigned long long value = pwdIndex;
 	for(int x = l-1; x >= 0; x--)
 	{
 		int charIndex = value % alphabetLen;
-		value = (long)value / alphabetLen;
+		value = (unsigned long long)value / alphabetLen;
 		pwd[x] = alphabet[charIndex];
 	}
 	pwd[l] = '\0';
@@ -62,7 +62,7 @@ void R(char *out, char *pwd, int k)
 	//Transform binary into a integer (long long)
 	unsigned __int128 hashValue = *((unsigned __int128*)out); 
 	hashValue += k;
-	int newPwdValue = hashValue % universe;
+	unsigned long long newPwdValue = hashValue % universe;
 	getPwd(pwd, newPwdValue);
 }
 
@@ -127,7 +127,7 @@ int main(int argc, char *argv[])
 		//Generate Seed		
 		while (1)
 		{	
-			int seed = rand() % (long)universe;
+			unsigned long long seed = rand() % (unsigned long long)universe;
 			getPwd(pwd, seed);
 			if(!ht_get(ht, pwd))
 				break;
